+++ src/types/sticker_set.rs
@@ -1,6 +1,8 @@
+use std::ops::Deref;
+
 use serde::{Deserialize, Serialize};
 
-use crate::types::{PhotoSize, Sticker};
+use crate::types::{PhotoSize, Sticker, StickerKind};
 
 /// This object represents a sticker set.
 ///
@@ -13,15 +15,8 @@ pub struct StickerSet {
     /// Sticker set title.
     pub title: String,
 
-    /// `true`, if the sticker set contains [animated stickers].
-    ///
-    /// [animates stickers]: https://telegram.org/blog/animated-stickers
-    pub is_animated: bool,
-
-    /// `true`, if the sticker is a [video sticker].
-    ///
-    /// [video sticker]: https://telegram.org/blog/video-stickers-better-reactions
-    pub is_video: bool,
+    /// Sticker kind shared by all stickers in this set.
+    pub kind: StickerKind,
 
     /// `true`, if the sticker set contains masks.
     pub contains_masks: bool,
@@ -31,4 +26,22 @@ pub struct StickerSet {
 
     /// Sticker set thumbnail in the .WEBP or .TGS format.
     pub thumb: Option<PhotoSize>,
+}
+
+/// This allows calling [`StickerKind`]'s methods directly on [`StickerSet`].
+///
+/// ```no_run
+/// use teloxide_core::types::StickerSet;
+///
+/// let sticker: StickerSet = todo!();
+///
+/// let _ = sticker.is_video();
+/// let _ = sticker.kind.is_video();
+/// ```
+impl Deref for StickerSet {
+    type Target = StickerKind;
+
+    fn deref(&self) -> &Self::Target {
+        &self.kind
+    }
 }